requirements.txt
flask
google-genai
pillow
python-dotenv
pyotp
